requirements.txt
#### Bot lib ####
git+https://github.com/HitsukiNetwork/Hitsuki-PTB.git
telethon>=1.17.5
pyrogram>=1.0.7
#################
aiohttp>=3.7.2
bleach>=3.2.1
emoji>=0.6.0
future>=0.18.2
googletrans>=3.0.0
markdown2>=2.3.10
Pillow>=8.0.1
PyYAML>=5.3.1
requests>=2.24.0
spamwatch>=0.3.0
sqlalchemy>=1.3.20
tldextract>=3.0.2
wikipedia>=1.4.0
covid>=2.4.0
bs4>=0.0.1
jikanpy>=4.2.2
python-rapidjson>=0.9.3

beautifulsoup4>=4.9.3
regex>=2020.10.28